requirements.txt
autopep8==1.7.0
python-dotenv==0.21.0
python-multipart==0.0.5
pandas==1.5.3
fastapi==0.85.0
uvicorn==0.18.3
torch==1.11.0
torchvision==0.12.0